refactor(prompts): move continuation prompt builder into RectifierPromptBuilder
buildReviewRectificationContinuation was living in src/pipeline/stages/
which violates the project convention that all prompt-building logic belongs
in src/prompts/builders/. Move it as a static method on RectifierPromptBuilder
and relocate CONTRADICTION_ESCAPE_HATCH there as the canonical definition.

- Add RectifierPromptBuilder.continuation() static method (PROMPT-001)
- Add CONTRADICTION_ESCAPE_HATCH export to rectifier-builder.ts + prompts barrel
- autofix-prompts.ts imports the constant from ../../prompts instead of defining it
- autofix.ts calls RectifierPromptBuilder.continuation() via ../../prompts barrel
- Delete src/pipeline/stages/autofix-continuation.ts
- Move tests: test/unit/pipeline/stages/autofix-continuation.test.ts
            → test/unit/prompts/builders/rectifier-builder.test.ts
